move App component into correct folder
