+++ components/App.jsx
@@ -1,4 +1,4 @@
-import TempoDisplay from '../components/TempoDisplay';
+import TempoDisplay from './TempoDisplay';
 
 export default function App() {
   return <TempoDisplay tempo={96} />;
